Operators unification
