+++ src/codegen/runner/sandbox/runner.py
@@ -3,7 +3,7 @@ import sys
 
 from git import Commit as GitCommit
 
-from codegen.git.repo_operator.remote_repo_operator import RemoteRepoOperator
+from codegen.git.repo_operator.repo_operator import RepoOperator
 from codegen.git.schemas.repo_config import RepoConfig
 from codegen.runner.models.apis import CreateBranchRequest, CreateBranchResponse, GetDiffRequest, GetDiffResponse
 from codegen.runner.sandbox.executor import SandboxExecutor
@@ -14,6 +14,7 @@ from codegen.sdk.secrets import Secrets
 from codegen.shared.compilation.string_to_code import create_execute_function_from_codeblock
 from codegen.shared.configs.session_configs import config
 from codegen.shared.performance.stopwatch_utils import stopwatch
+from codegen.git.schemas.enums import SetupOption
 
 logger = logging.getLogger(__name__)
 
@@ -24,7 +25,7 @@ class SandboxRunner:
     # =====[ __init__ instance attributes ]=====
     repo: RepoConfig
     commit: GitCommit
-    op: RemoteRepoOperator | None
+    op: RepoOperator | None
 
     # =====[ computed instance attributes ]=====
     codebase: CodebaseType
@@ -32,7 +33,7 @@ class SandboxRunner:
 
     def __init__(self, repo_config: RepoConfig, access_token: str) -> None:
         self.repo = repo_config
-        self.op = RemoteRepoOperator(repo_config=self.repo, access_token=access_token)
+        self.op = RepoOperator(repo_config=self.repo, access_token=access_token, setup_option=SetupOption.PULL_OR_CLONE)
         self.commit = self.op.git_cli.head.commit
 
     async def warmup(self) -> None:
@@ -56,7 +57,7 @@ class SandboxRunner:
 
         At the start of every job the runner should be in the following state:
         - Codebase is checked out to the pinned commit (i.e. self.commit)
-        - Codebase LRP (LocalRepoOperator) has only the origin remote and no branches
+        - Codebase RP (RepoOperator) has only the origin remote and no branches
 
         This method puts the runner in the above state and should be called at the end of every job.
         """
