+++ src/codegen/sdk/codebase/flagging/groupers/app_grouper.py
@@ -1,6 +1,6 @@
 import logging
 
-from codegen.git.repo_operator.remote_repo_operator import RemoteRepoOperator
+from codegen.git.repo_operator.repo_operator import RepoOperator
 from codegen.sdk.codebase.flagging.code_flag import CodeFlag
 from codegen.sdk.codebase.flagging.group import Group
 from codegen.sdk.codebase.flagging.groupers.base_grouper import BaseGrouper
@@ -17,7 +17,7 @@ class AppGrouper(BaseGrouper):
     type: GroupBy = GroupBy.APP
 
     @staticmethod
-    def create_all_groups(flags: list[CodeFlag], repo_operator: RemoteRepoOperator | None = None) -> list[Group]:
+    def create_all_groups(flags: list[CodeFlag], repo_operator: RepoOperator | None = None) -> list[Group]:
         unique_apps = list({"/".join(flag.filepath.split("/")[:3]) for flag in flags})
         groups = []
         for idx, app in enumerate(unique_apps):
@@ -27,7 +27,7 @@ class AppGrouper(BaseGrouper):
         return groups
 
     @staticmethod
-    def create_single_group(flags: list[CodeFlag], segment: str, repo_operator: RemoteRepoOperator | None = None) -> Group:
+    def create_single_group(flags: list[CodeFlag], segment: str, repo_operator: RepoOperator | None = None) -> Group:
         segment_flags = [f for f in flags if f.filepath.startswith(segment)]
         if len(segment_flags) == 0:
             logger.warning(f"🤷‍♀️ No flags found for APP segment: {segment}")
