+++ src/codegen/sdk/codebase/flagging/groupers/base_grouper.py
@@ -1,4 +1,4 @@
-from codegen.git.repo_operator.remote_repo_operator import RemoteRepoOperator
+from codegen.git.repo_operator.repo_operator import RepoOperator
 from codegen.sdk.codebase.flagging.code_flag import CodeFlag
 from codegen.sdk.codebase.flagging.group import Group
 from codegen.sdk.codebase.flagging.groupers.enums import GroupBy
@@ -18,12 +18,12 @@ class BaseGrouper:
             raise ValueError(msg)
 
     @staticmethod
-    def create_all_groups(flags: list[CodeFlag], repo_operator: RemoteRepoOperator | None = None) -> list[Group]:
+    def create_all_groups(flags: list[CodeFlag], repo_operator: RepoOperator | None = None) -> list[Group]:
         msg = "Must implement create_all_groups in BaseGrouper"
         raise NotImplementedError(msg)
 
     @staticmethod
-    def create_single_group(flags: list[CodeFlag], segment: str, repo_operator: RemoteRepoOperator | None = None) -> Group:
+    def create_single_group(flags: list[CodeFlag], segment: str, repo_operator: RepoOperator | None = None) -> Group:
         """TODO: handle the case when 0 flags are passed in"""
         msg = "Must implement create_single_group in BaseGrouper"
         raise NotImplementedError(msg)
