+++ src/codegen/sdk/codebase/flagging/groupers/file_grouper.py
@@ -1,6 +1,6 @@
 import logging
 
-from codegen.git.repo_operator.remote_repo_operator import RemoteRepoOperator
+from codegen.git.repo_operator.repo_operator import RepoOperator
 from codegen.sdk.codebase.flagging.code_flag import CodeFlag
 from codegen.sdk.codebase.flagging.group import Group
 from codegen.sdk.codebase.flagging.groupers.base_grouper import BaseGrouper
@@ -17,7 +17,7 @@ class FileGrouper(BaseGrouper):
     type: GroupBy = GroupBy.FILE
 
     @staticmethod
-    def create_all_groups(flags: list[CodeFlag], repo_operator: RemoteRepoOperator | None = None) -> list[Group]:
+    def create_all_groups(flags: list[CodeFlag], repo_operator: RepoOperator | None = None) -> list[Group]:
         groups = []
         filenames = sorted(list({f.filepath for f in flags}))
         for idx, filename in enumerate(filenames):
@@ -26,7 +26,7 @@ class FileGrouper(BaseGrouper):
         return groups
 
     @staticmethod
-    def create_single_group(flags: list[CodeFlag], segment: str, repo_operator: RemoteRepoOperator | None = None) -> Group:
+    def create_single_group(flags: list[CodeFlag], segment: str, repo_operator: RepoOperator | None = None) -> Group:
         segment_flags = [flag for flag in flags if flag.filepath == segment]
         if len(segment_flags) == 0:
             logger.warning(f"🤷‍♀️ No flags found for FILE segment: {segment}")
