+++ tests/integration/codegen/runner/test_create_branch.py
@@ -4,7 +4,7 @@ from http import HTTPStatus
 import pytest
 
 from codegen.git.clients.git_repo_client import GitRepoClient
-from codegen.git.repo_operator.remote_repo_operator import RemoteRepoOperator
+from codegen.git.repo_operator.repo_operator import RepoOperator
 from codegen.runner.clients.sandbox_client import SandboxClient
 from codegen.runner.models.apis import BRANCH_ENDPOINT, CreateBranchRequest, CreateBranchResponse
 from codegen.runner.models.codemod import BranchConfig, Codemod, GroupingConfig
@@ -12,7 +12,7 @@ from codegen.runner.models.codemod import BranchConfig, Codemod, GroupingConfig
 
 @pytest.mark.asyncio
 @pytest.mark.timeout(60)
-async def test_create_branch(sandbox_client: SandboxClient, git_repo_client: GitRepoClient, op: RemoteRepoOperator):
+async def test_create_branch(sandbox_client: SandboxClient, git_repo_client: GitRepoClient, op: RepoOperator):
     # set-up
     codemod_source = """
 for file in codebase.files:
